requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pandas
torch
sentence-transformers
scikit-learn


